+++ src/test/java/redis/embedded/RedisClientTest.java
@@ -22,13 +22,13 @@ import java.util.regex.Pattern;
 public class RedisClientTest {
 
     private RedisClient redisClient;
-    private int port;
+    private Integer port;
     private String host;
 
     @BeforeEach
     public void setUp() {
         host = CommonConstant.DEFAULT_REDIS_HOST;
-        port = RandomUtils.secure().randomInt(10001, 11000);
+        port = (Integer) RandomUtils.secure().randomInt(10001, 11000);
     }
 
     @Test
